Change default value for SitePolicy options to empty string
Set the default value of sitepolicymanager, etc. to be the empty
string so that bro will no longer load local-manager.bro, etc. by
default.  In order to make this work, I added two new options:
sitepolicyproxy and sitepolicylogger.  With all of these changes
it is now possible to remove the local-manager.bro, etc. scripts
(which are empty by default) without breaking anything.

Also added a warning message if a user tries to use SitePolicyStandalone
in their broctl.cfg.